src/main/java/adn/oscar/reservascanchas/dominio/servicio/cliente/ServicioActualizarCliente.java
package adn.oscar.reservascanchas.dominio.servicio.cliente;

import adn.oscar.reservascanchas.dominio.modelo.Cliente;
import adn.oscar.reservascanchas.dominio.repositorio.RepositorioCliente;

public class ServicioActualizarCliente {

    private final RepositorioCliente repositorioCliente;

    public ServicioActualizarCliente(RepositorioCliente repositorioCliente) {
        this.repositorioCliente = repositorioCliente;
    }

    public Cliente ejecutar(Cliente cliente) {
        return this.repositorioCliente.actualizar(cliente);
    }
}
